refactor(tests): move TestMaskROIHelpers to tests/detection/utils/test_masks.py
Mirrors source location of _mask_to_roi, _compact_masks_to_roi,
_masks_to_roi (detection/utils/masks.py). Removes dead try/except
import fallback from test_core.py.

---
