+++ tests/annotators/test_core.py
@@ -40,19 +40,6 @@ from supervision.draw.color import Color
 from supervision.geometry.core import Position
 from tests.helpers import _create_detections, assert_image_mostly_same
 
-try:
-    from supervision.detection.utils.masks import (
-        _compact_masks_to_roi,
-        _mask_to_roi,
-        _masks_to_roi,
-    )
-except ImportError:
-    from supervision.annotators.core import (  # type: ignore[no-redef]
-        _compact_masks_to_roi,
-        _mask_to_roi,
-        _masks_to_roi,
-    )
-
 
 @pytest.fixture
 def test_image() -> np.ndarray:
@@ -412,104 +399,6 @@ class TestMaskAnnotator:
             result_roi[roi_mask].astype(np.int16) - ref_blend[roi_mask].astype(np.int16)
         )
         assert np.all(diff <= 1), f"Max pixel diff: {diff.max()}"
-
-
-class TestMaskROIHelpers:
-    """Tests for _mask_to_roi, _compact_masks_to_roi, _masks_to_roi helpers."""
-
-    def test_mask_to_roi_all_false_returns_none(self):
-        """All-false mask should return None."""
-        mask = np.zeros((10, 15), dtype=bool)
-        assert _mask_to_roi(mask) is None
-
-    def test_mask_to_roi_single_pixel_exclusive_bounds(self):
-        """Single true pixel at (row=3, col=5) gives bounds (5, 3, 6, 4)."""
-        mask = np.zeros((10, 15), dtype=bool)
-        mask[3, 5] = True
-        assert _mask_to_roi(mask) == (5, 3, 6, 4)
-
-    def test_mask_to_roi_boundary_row_col_zero(self):
-        """True pixel at top-left boundary should give (0, 0, 1, 1)."""
-        mask = np.zeros((10, 15), dtype=bool)
-        mask[0, 0] = True
-        assert _mask_to_roi(mask) == (0, 0, 1, 1)
-
-    def test_mask_to_roi_full_image(self):
-        """Full-image true mask should span the entire array."""
-        h, w = 8, 12
-        mask = np.ones((h, w), dtype=bool)
-        assert _mask_to_roi(mask) == (0, 0, w, h)
-
-    def test_mask_to_roi_region(self):
-        """Region [10:20, 15:25] in 80x90 mask gives exclusive bounds (15,10,25,20)."""
-        mask = np.zeros((80, 90), dtype=bool)
-        mask[10:20, 15:25] = True
-        assert _mask_to_roi(mask) == (15, 10, 25, 20)
-
-    def test_compact_masks_to_roi_empty_returns_none(self):
-        """Zero-length CompactMask should return None."""
-        masks = np.zeros((0, 10, 10), dtype=bool)
-        xyxy = np.empty((0, 4), dtype=np.float32)
-        cm = CompactMask.from_dense(masks, xyxy, image_shape=(10, 10))
-        assert _compact_masks_to_roi(cm, (10, 10)) is None
-
-    def test_compact_masks_to_roi_single_detection_coordinates(self):
-        """Single compact mask with known crop should give correct exclusive bounds."""
-        masks = np.zeros((1, 20, 20), dtype=bool)
-        masks[0, 5:10, 3:8] = True
-        xyxy = np.array([[3.0, 5.0, 7.0, 9.0]], dtype=np.float32)
-        cm = CompactMask.from_dense(masks, xyxy, image_shape=(20, 20))
-        result = _compact_masks_to_roi(cm, (20, 20))
-        assert result is not None
-        x1, y1, x2, y2 = result
-        # Bounding-box crop defines the extent; true pixels lie within [3:8, 5:10]
-        assert x1 <= 3
-        assert y1 <= 5
-        assert x2 >= 8
-        assert y2 >= 10
-
-    def test_compact_masks_to_roi_two_detections_union(self):
-        """Two disjoint compact masks union should span both regions."""
-        masks = np.zeros((2, 30, 40), dtype=bool)
-        masks[0, 2:8, 1:6] = True
-        masks[1, 15:25, 20:35] = True
-        xyxy = np.array(
-            [[1.0, 2.0, 5.0, 7.0], [20.0, 15.0, 34.0, 24.0]], dtype=np.float32
-        )
-        cm = CompactMask.from_dense(masks, xyxy, image_shape=(30, 40))
-        result = _compact_masks_to_roi(cm, (30, 40))
-        assert result is not None
-        x1, y1, x2, y2 = result
-        # Union must contain both regions
-        assert x1 <= 1
-        assert y1 <= 2
-        assert x2 >= 35
-        assert y2 >= 25
-
-    def test_masks_to_roi_empty_array_returns_none(self):
-        """Zero-element dense mask array (N=0) should return None."""
-        masks = np.zeros((0, 10, 10), dtype=bool)
-        assert _masks_to_roi(masks, (10, 10)) is None
-
-    def test_masks_to_roi_2d_array(self):
-        """2D boolean array (single mask) should work as union mask."""
-        mask = np.zeros((10, 15), dtype=bool)
-        mask[3:6, 7:11] = True
-        assert _masks_to_roi(mask, (10, 15)) == (7, 3, 11, 6)
-
-    def test_masks_to_roi_dense_with_xyxy_uses_box_union(self):
-        """Dense path with xyxy should return union of boxes (O(N) path)."""
-        masks = np.zeros((2, 30, 40), dtype=bool)
-        masks[0, 5:10, 5:10] = True
-        masks[1, 20:25, 25:30] = True
-        xyxy = np.array([[5.0, 5.0, 9.0, 9.0], [25.0, 20.0, 29.0, 24.0]])
-        result = _masks_to_roi(masks, (30, 40), xyxy)
-        assert result is not None
-        x1, y1, x2, y2 = result
-        assert x1 <= 5
-        assert y1 <= 5
-        assert x2 >= 30
-        assert y2 >= 25
 
 
 class TestPolygonAnnotator:
